begin working backwards from config file clarifying names
* free from other context names should be as specific
as possible, so zainod-* where the indexer must be
zainod, e.g. in the config file for zainod.
* in zainod and other config names should include
purpose and protocol e.g. controlplane_rpc-, or
p2p_grpc-
* in a given config which component (validator,
indexer, client) purpose (control, p2p) and protocol
(gRPC, RPC) is a appropriate, note that self is
implied in those cases